+++ client/package.json
@@ -10,6 +10,7 @@
     "axios": "^1.1.3",
     "gapi-script": "^1.2.0",
     "react": "^18.2.0",
+    "react-aws-s3": "^1.5.0",
     "react-datepicker": "^4.8.0",
     "react-dom": "^18.2.0",
     "react-google-login": "^5.2.2",
